publish: version 0.6.3
