+++ package.json
@@ -7,7 +7,7 @@
     "url": "git+https://github.com/helpwave/hightide.git"
   },
   "license": "MPL-2.0",
-  "version": "0.6.2",
+  "version": "0.6.3",
   "files": [
     "dist"
   ],
