+++ src/components/layout/dialog/LanguageDialog.tsx
@@ -3,7 +3,7 @@ import type { DialogProps } from '@/src/components/layout/dialog/Dialog'
 import { Dialog } from '@/src/components/layout/dialog/Dialog'
 import { LocalizationUtil } from '@/src/i18n/util'
 import { useLocale } from '@/src/contexts/LocaleContext'
-import { Select, SelectOption } from '@/src/components/user-interaction/Select'
+import { Select, SelectOption } from '@/src/components/user-interaction/select'
 import { Button } from '@/src/components/user-interaction/Button'
 import { useHightideTranslation } from '@/src/i18n/useHightideTranslation'
 import type { HightideTranslationLocales } from '@/src/i18n/translations'
